requirements.txt
# Core Python optimization
uvloop==0.19.0
orjson==3.9.10
pysimdjson==5.0.2

# Exchange connectivity & execution
ccxt==4.1.80
aiohttp==3.9.1
asyncio==3.4.3
websockets==12.0

# Data structures & IPC
pyzmq==25.1.2
msgpack==1.0.7
numpy==1.26.2
pandas==2.1.4

# Database connectors
psycopg2-binary==2.9.9
clickhouse-driver==0.2.6
sqlalchemy==2.0.23
alembic==1.13.0

# Technical analysis
ta==0.11.0
TA-Lib==0.4.28

# Machine Learning
scikit-learn==1.3.2
lightgbm==4.1.0
onnxruntime==1.16.3
shap==0.43.0

# Deep Learning (optional, for L4 LLM)
torch==2.1.1
transformers==4.36.0
bitsandbytes==0.41.3
langchain==0.1.0
anthropic==0.8.0

# Reinforcement Learning
ray[rllib]==2.9.0

# Backtesting & Research
backtrader==1.9.78.123

# Monitoring & Observability
prometheus-client==0.19.0
statsd==4.0.1
psutil==5.9.6

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
tenacity==8.2.3
keyring==24.3.0
cryptography==41.0.7

# Time synchronization
ntplib==0.4.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-timeout==2.2.0

# Development tools
black==23.12.1
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

# API frameworks (for risk_proxy)
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3

# Chaos engineering
toxiproxy-py==0.6.0

# News & sentiment
beautifulsoup4==4.12.2
feedparser==6.0.10
vaderSentiment==3.3.2

# NLP
nltk==3.8.1
spacy==3.7.2

# DEX support (future)
solana==0.30.2
anchorpy==0.18.0

# Configuration management
python-json-logger==2.0.7
structlog==23.2.0

# Performance profiling
py-spy==0.3.14
memory-profiler==0.61.0
